+++ src/views/Home.jsx
@@ -6,6 +6,7 @@ import {
 } from '@material-ui/core';
 import CategoriesView from '../components/CategoriesView';
 import NavBar from '../components/NavBar';
+import EventList from '../components/EventList';
 
 class Home extends React.Component {
 
@@ -38,6 +39,9 @@ class Home extends React.Component {
                 <div className="categories">
                     <Typography variant="h4" className="title">Kategorien</Typography>
                     <CategoriesView onClick={category => history.push("/category/" + category)} />
+                
+                    <Typography variant="h4" className="title">Vorgeschlagen</Typography>
+                    <EventList categoryId="more" limit={5} />
                 </div>
                 {/* Under construction stuff */}
                 <div style={{ height: '150px' }}></div>
